Move recurrence in concern
